requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
insightface>=0.7.3
onnxruntime>=1.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
safetensors>=0.3.0
